[nvim] rename config files into snake_case for consistency
