+++ nvim/.config/nvim/lua/user/lsp/init.lua
@@ -1,3 +1,3 @@
 require('user.lsp.handlers').setup()
-require('user.lsp.lsp-installer')
+require('user.lsp.lsp_installer')
