+++ nvim/.config/nvim/lua/user/plugins_configs/init.lua
@@ -4,11 +4,12 @@ require('user.plugins_configs.misc')
 require('user.plugins_configs.treesitter')
 -- temporarily swithced neoscroll off
 -- require('user.plugins_configs.neoscroll')
-require('user.plugins_configs.nvim-tree')
+require('user.plugins_configs.nvim_tree')
 require('user.plugins_configs.lualine')
 require('user.plugins_configs.material')
-require('user.plugins_configs.nvim-ufo')
+require('user.plugins_configs.nvim_ufo')
 require('user.plugins_configs.floaterm')
+-- require('user.plugins_configs.solar_client')
 -- winbar requires NeoVim nightly, aka 0.9.x
 -- require('user.plugins.winbar')
