+++ nvim/.config/nvim/lua/user/plugins_configs/nvim_ufo.lua
@@ -2,9 +2,9 @@ local utils = require('user.utils')
 local ufo = utils.safe_require('ufo')
 
 ufo.setup({
-    provider_selector = function(bufnr, filetype, buftype)
-        return {'treesitter', 'indent'}
-    end
+  provider_selector = function(bufnr, filetype, buftype)
+    return {'treesitter', 'indent'}
+  end
 })
 
 -- NOTE: if this set to e.g. 0, on opening file
